Lightos - add lightos new volume states
Lighbits volumes have a state field indicating the state of the volume
as per actions that were invoked for this
volume (create, delete, attach...).

New states were recently added to volume:
- Migrating - specifying the lightbits dynamic rebalance logic,
  is currently migrating one of the replicas of this volume
  within the cluster to better balance the cluster.
- Rollback - Rollback the data of the volume to a previously
  created snapshot of this volume.

Migrating is identical to Avaialble with respect to most operations.
Rollback has various restriction (more similiar to Updating state)
on concurrent operations while in this state.
this patch adds handling of volume actions additional
considerations of the new states.

Signed-off-by: Yuval Brave  <yuval@lightbitslabs.com>
Change-Id: Icf873eedde145de0a7e2696036bf30ed3ca36a8f